RDF store overhaul
Changed the basic Store type to an abstract type, removing its ability to be interpreted as a mutable type. The mutable variant and a readonly version is now the most basic versions of the RDF store. A new version, the ConcurrentMutableStore, has also been added, using platform-specific locking mechanisms for reading and writing safely across multiple threads in a multithreaded environment
Because of the requirement of thread-safe element iteration, the regular iterator implementation has been dropped in favour of a manual `forEach` implementation that iterates whilst holding the read lock

Adapted most uses of RDF collections in (public) APIs to use the basic Store type as its foundation, relying on custom extension function implementations acting as an Iterable type

Changed `Iterator<Quad>.consume` to `.toStore`, making it more clear
Changed the semantics of the `replace` method to remove its key-value pair from the map if the value is set to `null`

WIP: RWLock is not implemented yet for native targets! The Linux target can use https://man7.org/linux/man-pages/man3/pthread_rwlock_rdlock.3p.html