Store API overhaul
Reworked the RDF Store type: the base Store is now an interface, with a specialised `iter` method to allow more direct filtering when obtaining quads inside the store, which the used store implementation can use to facilitate more efficient data retrieval.
The basic Store type is no longer mutable. Instead, a type implementing the MutableStore interface is required. Furthermore, MutableStore types are not observable by default, and is further fragmented into a ObservableStore interface. As observing immutable stores is not useful, ObservableStores are automatically required to be mutable as well
All interface types have concrete in-memory implementations available, and are accessible through new factory functions named after the interface they implement
A new, but currently unused, IndexedStore has also been added, which facilitates the Store.iter method more efficiently
The MutableStore now also properly implements the MutableCollection interface, allowing MutableStore implementations to be used in a set of standard library functions

The type overhaul makes it possible for different store implementations to be considered equal, a single Store interface to be sufficient in scenarios where readonly access is required, a single MutableStore interface is usable for readwrite access, and ObservableStore interface is usable for readwrite with observations is required, regardless of their implementations

Applied the various Store type changes throughout all other modules
Added the graph term to the string representation of quads
Added the graph term to the string representation of stores