Added `CompleteHashMappingArray` and `MappingArrayFlags`
Added a new `MappingArray` implementation, similar in behaviour as the original `MultiHashMappingArray` implementation, but throwing an exception when attempting to use a partial index instead of falling back to a linear lookup
Added a new hint, `MappingArrayHint`, used to configure the required features of the chosen implementation when creating a new array instance. Currently, only the need for partial index access is available
Added the ability to reindex mapping arrays using the new hint
Renamed `MutableJoinState::rehash` to `MutableJoinState::reindex` for consistency, which now also uses the new hint type
Removed `StatelessJoinTree`, replacing it with a `SingleItemJoinTree` when exactly one join state is part of a larger structure (e.g. query body), expecting a `JoinTree` implementor
Changed `MultiHashMappingArray` to expect at least two bindings to index on
Changed `SimpleMappingArray`'s failure exception to `NoSuchElementException` for API consistency