`DynamicJoinTree::Node::Connected` data deletion fix
Fixed deletions in queries affecting multiple TPs at the same time while one TP is part of a `Node::Connected` instance, resulting in a join result to be created by joining onto itself. The implementation's `buf` caching the results between the two child nodes `left` and `right` contains the to-be-deleted result at the time of `join` being called, yielding an additional, incorrect, result. Now, the same strategy found in `TriplePatternState::ArrayBackedPatternState` is applied, which removes exactly one match from the to-be-joined stream that corresponds to change that the data deletion itself causes, preventing such an erroneous join result from existing.
Fixes the output validation tests